custom hook functionality added.
