+++ pages/_app.jsx
@@ -1,13 +1,12 @@
 import React from 'react';
-import { useRouter } from 'next/router';
 import { BaseLayout } from '@components/base_layout';
 import 'bootstrap/dist/css/bootstrap.min.css';
+import { useIsHomePage } from '@utils/hooks';
 import '@styles/globals.scss';
 import Home from '.';
 
 function App({ Component, pageProps }) {
-    const router = useRouter();
-    const isHomePage = router.pathname === '/';
+    const isHomePage = useIsHomePage();
 
     const getLayout = Component.getLayout || ((page) => page);
 
@@ -16,10 +15,10 @@ function App({ Component, pageProps }) {
         require('bootstrap/dist/js/bootstrap.bundle.min.js');
     }, []);
 
-    return isHomePage ? (
-        <Home />
-    ) : (
-        <BaseLayout>{getLayout(<Component {...pageProps} />)}</BaseLayout>
+    return (
+        <BaseLayout hideSidebar={isHomePage}>
+            {getLayout(<Component {...pageProps} />)}
+        </BaseLayout>
     );
 }
