+++ pages/index.jsx
@@ -20,19 +20,6 @@ export default function Home() {
 
     return (
         <>
-            <Head>
-                <link rel="preconnect" href="https://fonts.googleapis.com" />
-                <link
-                    rel="preconnect"
-                    href="https://fonts.gstatic.com"
-                    crossorigin
-                />
-                <link
-                    href="https://fonts.googleapis.com/css2?family=Inter:wght@200;300;500&family=Shadows+Into+Light&display=swap"
-                    rel="stylesheet"
-                />
-            </Head>
-            <Header />
             <div className={css.hero}>
                 <h1>Theology101.church</h1>
             </div>
