+++ utils/hooks.js
@@ -1,0 +1,6 @@
+import { useRouter } from 'next/router';
+
+export function useIsHomePage() {
+    const router = useRouter();
+    return router.pathname === '/';
+}
